v0.1.20 release handler
